feature(FrondEnd): Backend API Rest methods consumption
