+++ src/Components/Molecules/Sections/Section1.tsx
@@ -1,0 +1,8 @@
+export default function Section1() {
+    return(
+        <div className="w-full xl:w-8/12 mb-12 xl:mb-0 px-4 min-h-5">
+          {/* <CardLineChart /> */}
+          Section 1 tables
+        </div>
+    )
+}
