+++ src/Components/Molecules/Sections/Section2.tsx
@@ -1,0 +1,10 @@
+export default function Section2() {
+    return(
+        <>
+        <div className="w-full xl:w-4/12 px-4 min-h-5">
+          {/* <CardBarChart /> */}
+          Section 2 Line Charts
+        </div>
+        </>
+    )
+}
