+++ src/Components/Molecules/Sections/Section3.tsx
@@ -1,0 +1,8 @@
+export default function Section3() {
+    return (
+        <div className="w-full xl:w-8/12 mb-12 xl:mb-0 px-4 min-h-5">
+          {/* <CardPageVisits /> */}
+          Section 3 Bar Charts
+        </div>
+    )
+}
