+++ src/Components/Molecules/Sections/Section4.tsx
@@ -1,0 +1,8 @@
+export default function Section4() {
+    return (
+        <div className="w-full xl:w-4/12 px-4 min-h-5">
+          {/* <CardSocialTraffic /> */}
+          Section 4 High Scores
+        </div>
+    )
+}
