+++ src/app/Auth/Login/page.tsx
@@ -78,19 +78,6 @@ export default function Login() {
                       placeholder="Password"
                     />
                   </div>
-                  <div>
-                    <label className="inline-flex items-center cursor-pointer">
-                      <input
-                        id="customCheckLogin"
-                        type="checkbox"
-                        className="form-checkbox border-0 rounded text-gray-700 ml-1 w-5 h-5 ease-linear transition-all duration-150"
-                      />
-                      <span className="ml-2 text-sm font-semibold text-gray-600">
-                        Remember me
-                      </span>
-                    </label>
-                  </div>
-
                   <div className="text-center mt-6">
                     <button
                       className="bg-gray-800 text-white active:bg-gray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full ease-linear transition-all duration-150"
@@ -115,7 +102,7 @@ export default function Login() {
               <div className="w-1/2 text-right">
                 <Link 
                 className="text-gray-200"
-                href="/auth/Register">
+                href="/Auth/Register">
                   {/* <a href="#pablo" className="text-gray-200"> */}
                     <small>Create new account</small>
                  {/*  </a> */}
