+++ src/app/Auth/Register/page.tsx
@@ -3,9 +3,24 @@
 import React from "react";
 import FooterAuth from "game-score-frontend/Components/Footers/FooterAuth";
 import Navbar from "game-score-frontend/Components/Navbars/AuthNavbar";
-import Link from "next/link";
+import RegisterForm from "game-score-frontend/Components/Forms/RegisterForm";
 
 export default function Register() {
+
+  interface ApiResponse {
+    id: string;
+    name: string;
+    username: string;
+    email: string;
+    // Agrega otros campos que pueda devolver la API si es necesario
+  }
+
+  const handleSuccess = (data: ApiResponse) => {
+    // Muestra los datos de usuario registrado en la consola
+    console.log("Registered user data:", data);
+    localStorage.setItem("user", JSON.stringify(data));
+  };
+
 
   return (
     <>
@@ -50,97 +65,7 @@ export default function Register() {
                 <div className="text-gray-400 text-center mb-3 font-bold">
                   <small>Or sign up with credentials</small>
                 </div>
-                <form>
-                  <div className="relative w-full mb-3">
-                    <label
-                      className="block uppercase text-gray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
-                    >
-                      Name
-                    </label>
-                    <input
-                      type="email"
-                      className="border-0 px-3 py-3 placeholder-gray-300 text-gray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Name"
-                      name="name"
-                    />
-                  </div>
-                  <div className="relative w-full mb-3">
-                    <label
-                      className="block uppercase text-gray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
-                    >
-                      UserName (NickName)
-                    </label>
-                    <input
-                      type="email"
-                      className="border-0 px-3 py-3 placeholder-gray-300 text-gray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="userName"
-                      name="username"
-                    />
-                  </div>
-
-                  <div className="relative w-full mb-3">
-                    <label
-                      className="block uppercase text-gray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
-                    >
-                      Email
-                    </label>
-                    <input
-                      type="email"
-                      className="border-0 px-3 py-3 placeholder-gray-300 text-gray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Email"
-                      name="email"
-                    />
-                  </div>
-
-                  <div className="relative w-full mb-3">
-                    <label
-                      className="block uppercase text-gray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
-                    >
-                      Password
-                    </label>
-                    <input
-                      type="password"
-                      className="border-0 px-3 py-3 placeholder-gray-300 text-gray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Password"
-                      name="password"
-                    />
-                  </div>
-
-                  <div>
-                    <label className="inline-flex items-center cursor-pointer">
-                      <input
-                        id="customCheckLogin"
-                        type="checkbox"
-                        className="form-checkbox border-0 rounded text-gray-700 ml-1 w-5 h-5 ease-linear transition-all duration-150"
-                      />
-                      <span className="ml-2 text-sm font-semibold text-gray-600">
-                        I agree with the{" "}
-                        <a
-                          href="#pablo"
-                          className="text-lightBlue-500"
-                          onClick={(e) => e.preventDefault()}
-                        >
-                          Privacy Policy
-                        </a>
-                      </span>
-                    </label>
-                  </div>
-
-                  <div className="text-center mt-6">
-                  <Link href="/Admin/Dashboard">
-                    <button
-                      className="bg-gray-800 text-white active:bg-gray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full ease-linear transition-all duration-150"
-                      type="button"
-                    >
-                      Create Account
-                    </button>
-                  </Link>
-                  </div>
-                </form>
+                <RegisterForm onSuccess={handleSuccess}/>
               </div>
             </div>
           </div>
