--- src/app/api/v1/users/scores/[id]/route.tsx
@@ -1,37 +1,0 @@
-import { NextResponse } from 'next/server';
-import data from 'game-score-frontend/data/fake_game_scores_data.json';
-
-export async function GET(request: Request, { params }: { params: { id: string } }) {
-  try {
-    const authHeader = request.headers.get('Authorization');
-    if (!authHeader || !authHeader.startsWith('Bearer ')) {
-      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
-    }
-
-    const token = authHeader.split(' ')[1];
-
-    if (token !== 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiIxMjNlNDU2Ny1lODliLTEyZDMtYTQ1Ni00MjY2MTQxNzQwMDAiLCJpYXQiOjE3Mjg4Njk0MTIsImV4cCI6MTcyODg4MzgxMn0.MaXY4Lr8jpOQ_v-Jf_iSKl9vVXLHhnRSh4t6SvktyFE') {
-      return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
-    }
-
-    const { id: playerId } = params;
-
-    if (!playerId) {
-      return NextResponse.json({ message: 'No playerId provided in the URL' }, { status: 400 });
-    }
-
-    // Filtrar los scores que coincidan con el playerId proporcionado
-    const playerScores = data.scores.filter(score => score.playerId === playerId);
-
-    if (playerScores.length === 0) {
-      return NextResponse.json({ message: `No scores found for playerId: ${playerId}` }, { status: 404 });
-    }
-
-    return NextResponse.json(playerScores, { status: 200 });
-  } catch (error) {
-    return NextResponse.json(
-      { message: 'Server error', error: (error as Error).message },
-      { status: 500 }
-    );
-  }
-}
